Separate logback conf for free/pro version. Also some refactoring
