+++ smartReceiptsFree/build-types/debug/assets/logback.xml
@@ -8,15 +8,22 @@
         </encoder>
     </appender>
 
-    <appender name="file" class="ch.qos.logback.core.FileAppender">
+    <appender name="file" class="ch.qos.logback.core.rolling.RollingFileAppender">
         <file>/data/data/wb.receipts/files/log.txt</file>
         <encoder>
             <pattern>[%date] &lt;%logger{20}&gt; %level: %msg%n</pattern>
         </encoder>
+        <rollingPolicy class="ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
+            <maxIndex>1</maxIndex>
+            <FileNamePattern>/data/data/wb.receipts/files/log%i.txt</FileNamePattern>
+        </rollingPolicy>
+        <triggeringPolicy class="ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
+            <MaxFileSize>1MB</MaxFileSize>
+        </triggeringPolicy>
     </appender>
 
     <root level="DEBUG">
-        <!--<appender-ref ref="logcat" />-->
+        <appender-ref ref="logcat" />
         <appender-ref ref="file" />
     </root>
 </configuration>
