+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/analytics/AnalyticsLogger.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import co.smartreceipts.android.analytics.events.DataPoint;
 import co.smartreceipts.android.analytics.events.Event;
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 
 public class AnalyticsLogger implements Analytics {
