+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/persistence/database/controllers/impl/ColumnTableController.java
@@ -19,7 +19,7 @@ import co.smartreceipts.android.persistence.database.controllers.alterations.Tab
 import co.smartreceipts.android.persistence.database.operations.DatabaseOperationMetadata;
 import co.smartreceipts.android.persistence.database.tables.AbstractColumnTable;
 import co.smartreceipts.android.utils.ListUtils;
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 import rx.Scheduler;
 import rx.Subscription;
 import rx.android.schedulers.AndroidSchedulers;
