+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/persistence/database/controllers/impl/TripForeignKeyAbstractTableController.java
@@ -14,7 +14,7 @@ import co.smartreceipts.android.persistence.database.tables.TripForeignKeyAbstra
 import co.smartreceipts.android.persistence.database.controllers.TableEventsListener;
 import co.smartreceipts.android.persistence.database.controllers.TripForeignKeyTableEventsListener;
 import co.smartreceipts.android.persistence.database.controllers.alterations.TableActionAlterations;
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 import rx.Observable;
 import rx.Scheduler;
 import rx.Subscription;
