+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/sync/manual/ManualBackupTask.java
@@ -12,7 +12,7 @@ import java.io.IOException;
 import co.smartreceipts.android.date.DateUtils;
 import co.smartreceipts.android.persistence.DatabaseHelper;
 import co.smartreceipts.android.persistence.PersistenceManager;
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 import rx.Observable;
 import rx.Scheduler;
 import rx.Subscriber;
