+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/sync/manual/ManualRestoreTask.java
@@ -16,7 +16,7 @@ import java.util.Map;
 import co.smartreceipts.android.persistence.DatabaseHelper;
 import co.smartreceipts.android.persistence.PersistenceManager;
 import co.smartreceipts.android.utils.ExceptionUtils;
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 import rx.Observable;
 import rx.Scheduler;
 import rx.Subscriber;
