+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/sync/network/AbstractNetworkProvider.java
@@ -12,7 +12,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import co.smartreceipts.android.utils.Logger;
+import co.smartreceipts.android.utils.log.Logger;
 
 abstract class AbstractNetworkProvider implements NetworkProvider {
