+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/utils/WBUncaughtExceptionHandler.java
@@ -9,6 +9,7 @@ import java.io.StringWriter;
 import java.lang.Thread.UncaughtExceptionHandler;
 
 import co.smartreceipts.android.SmartReceiptsApplication;
+import co.smartreceipts.android.utils.log.Logger;
 
 /**
  * This enables us to handle uncaught exceptions in a customizable manner. This is needed to fix a bug with Google Play
