+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/utils/log/LogConstants.java
@@ -1,0 +1,8 @@
+package co.smartreceipts.android.utils.log;
+
+public class LogConstants {
+
+    public static final String LOG_FILE_NAME_1 = "log.txt";
+    public static final String LOG_FILE_NAME_2 = "log1.txt";
+
+}
