+++ smartReceiptsPRO/build-types/release/assets/logback.xml
@@ -1,0 +1,29 @@
+<configuration>
+    <appender name="logcat" class="ch.qos.logback.classic.android.LogcatAppender">
+        <tagEncoder>
+            <pattern>%logger{12}</pattern>
+        </tagEncoder>
+        <encoder>
+            <pattern>[%date] &lt;%logger{20}&gt; %level: %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <appender name="file" class="ch.qos.logback.core.rolling.RollingFileAppender">
+        <file>/data/data/wb.receiptspro/files/log.txt</file>
+        <encoder>
+            <pattern>[%date] &lt;%logger{20}&gt; %level: %msg%n</pattern>
+        </encoder>
+        <rollingPolicy class="ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
+            <maxIndex>1</maxIndex>
+            <FileNamePattern>/data/data/wb.receiptspro/files/log%i.txt</FileNamePattern>
+        </rollingPolicy>
+        <triggeringPolicy class="ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
+            <MaxFileSize>1MB</MaxFileSize>
+        </triggeringPolicy>
+    </appender>
+
+    <root level="DEBUG">
+        <!--<appender-ref ref="logcat" />-->
+        <appender-ref ref="file" />
+    </root>
+</configuration>
